+++ java/src/main/java/com/angrysurfer/midi/model/Strike.java
@@ -9,6 +9,7 @@ import jakarta.persistence.*;
 
 import java.util.*;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 @Getter
 @Setter
@@ -19,31 +20,31 @@ public class Strike extends Player {
 
     static final Random rand = new Random();
 
-    public static int KICK = 36;
-    public static int SNARE = 37;
-    public static int CLOSED_HAT = 38;
-    public static int OPEN_HAT = 39;
+    public static long KICK = 36;
+    public static long SNARE = 37;
+    public static long CLOSED_HAT = 38;
+    public static long OPEN_HAT = 39;
 
     public static List<Integer> razParams = List.of(16, 17, 18, 19, 20, 21, 22, 23);
     public static List<Integer> closedHatParams = List.of(24, 25, 26, 27, 28, 29, 30, 31);
     public static List<Integer> kickParams = List.of(1, 2, 3, 4, 12, 13, 14, 15);
     public static List<Integer> snarePrams = List.of(16, 17, 18, 19, 20, 21, 22, 23);
 
-    private int ratchetCount = 0;
+    private Long ratchetCount = 0L;
     
-    private int ratchetInterval = 0;
+    private Long ratchetInterval = 0L;
 
     public Strike() {
         setNote(KICK);    
     }
 
-    public Strike(String name, Ticker ticker, MidiInstrument instrument, int note, List<Integer> allowedControlMessages) {
+    public Strike(String name, Ticker ticker, MidiInstrument instrument, long note, List<Integer> allowedControlMessages) {
         super(name, ticker, instrument, allowedControlMessages);
         setNote(note);
     }
 
-    public Strike(String name, Ticker ticker, MidiInstrument instrument, int note,
-            List<Integer> allowableControlMessages, int minVelocity, int maxVelocity) {
+    public Strike(String name, Ticker ticker, MidiInstrument instrument, long note,
+            List<Integer> allowableControlMessages, long minVelocity, long maxVelocity) {
         super(name, ticker, instrument, allowableControlMessages);
         setNote(note);
         setMinVelocity(minVelocity);
@@ -73,7 +74,7 @@ public class Strike extends Player {
             if (getSkipCycler().getLength() == 0 || getSkipCycler().atEnd()) {
                 // else handleSwing();
                 handleRachets();
-                drumNoteOn(getNote(), rand.nextInt(getMinVelocity(), getMaxVelocity()));
+                drumNoteOn(getNote(), rand.nextLong(getMinVelocity(), getMaxVelocity()));
             }
     
             // getSkipCycler().advance();
@@ -85,7 +86,7 @@ public class Strike extends Player {
     }
 
     private void handleRachets() {
-        IntStream.range(1, getRatchetCount() + 1).forEach(i -> {
+        LongStream.range(1, getRatchetCount() + 1).forEach(i -> {
             double base = getTicker().getTicksPerBeat() / getTicker().getBeatsPerBar(); 
             double offset = getTicker().getTick() + (base * i * getRatchetInterval());
             new Ratchet(this,(long) offset, getRatchetInterval(), 0);
